refactor(ui): extract field value-inputs into shared Fields module
Lift FormLayout's 23 fieldtype value-input components, the fieldType dispatch
registry, and fieldtypeToLanguage into a sibling src/components/Fields module so
both FormLayout and the upcoming ListView Filter consume them without reaching
across a module boundary (ADR-0004).

The field contract (FieldMeta, FieldComponentProps, FieldComponentEmits) and the
form-context injection keys (DocKey, ParentDocKey, UpdateKey) move into
Fields/types.ts; FormLayout/types.ts re-exports them so existing imports keep
working. Adds the ./fields export subpath.
