+++ ui/src/components/Fields/CheckField.vue
@@ -11,7 +11,7 @@
 <script setup lang="ts">
 import { computed } from "vue";
 import { Checkbox } from "frappe-ui";
-import type { FieldComponentEmits, FieldComponentProps } from "../types";
+import type { FieldComponentEmits, FieldComponentProps } from "./types";
 
 const props = defineProps<FieldComponentProps>();
 const emit = defineEmits<FieldComponentEmits>();
