+++ ui/src/components/Fields/DateField.vue
@@ -12,7 +12,7 @@
 <script setup lang="ts">
 import { computed } from "vue";
 import { DatePicker } from "frappe-ui";
-import type { FieldComponentEmits, FieldComponentProps } from "../types";
+import type { FieldComponentEmits, FieldComponentProps } from "./types";
 
 const props = defineProps<FieldComponentProps>();
 const emit = defineEmits<FieldComponentEmits>();
