+++ ui/src/components/Fields/DatetimeField.vue
@@ -12,7 +12,7 @@
 <script setup lang="ts">
 import { computed } from "vue";
 import { DateTimePicker } from "frappe-ui";
-import type { FieldComponentEmits, FieldComponentProps } from "../types";
+import type { FieldComponentEmits, FieldComponentProps } from "./types";
 
 const props = defineProps<FieldComponentProps>();
 const emit = defineEmits<FieldComponentEmits>();
