+++ ui/src/components/Fields/tests/fieldTypes.test.ts
@@ -1,29 +1,29 @@
 import { describe, expect, it } from "vitest";
 import { defineComponent } from "vue";
 import { getFieldComponent, registerFieldType } from "../fieldTypes";
-import AutocompleteField from "../fields/AutocompleteField.vue";
-import ButtonField from "../fields/ButtonField.vue";
-import CheckField from "../fields/CheckField.vue";
-import CodeEditorField from "../fields/CodeEditorField.vue";
-import DateField from "../fields/DateField.vue";
-import DatetimeField from "../fields/DatetimeField.vue";
-import DurationField from "../fields/DurationField.vue";
-import DynamicLinkField from "../fields/DynamicLinkField.vue";
-import AttachField from "../fields/AttachField.vue";
-import ImageField from "../fields/ImageField.vue";
-import GeolocationField from "../fields/GeolocationField.vue";
-import HeadingField from "../fields/HeadingField.vue";
-import HtmlField from "../fields/HtmlField.vue";
-import LinkField from "../fields/LinkField.vue";
-import NumberField from "../fields/NumberField.vue";
-import PasswordField from "../fields/PasswordField.vue";
-import PhoneField from "../fields/PhoneField.vue";
-import RatingField from "../fields/RatingField.vue";
-import SelectField from "../fields/SelectField.vue";
-import TableMultiSelectField from "../fields/TableMultiSelectField.vue";
-import TextField from "../fields/TextField.vue";
-import TextareaField from "../fields/TextareaField.vue";
-import TimeField from "../fields/TimeField.vue";
+import AutocompleteField from "../AutocompleteField.vue";
+import ButtonField from "../ButtonField.vue";
+import CheckField from "../CheckField.vue";
+import CodeEditorField from "../CodeEditorField.vue";
+import DateField from "../DateField.vue";
+import DatetimeField from "../DatetimeField.vue";
+import DurationField from "../DurationField.vue";
+import DynamicLinkField from "../DynamicLinkField.vue";
+import AttachField from "../AttachField.vue";
+import ImageField from "../ImageField.vue";
+import GeolocationField from "../GeolocationField.vue";
+import HeadingField from "../HeadingField.vue";
+import HtmlField from "../HtmlField.vue";
+import LinkField from "../LinkField.vue";
+import NumberField from "../NumberField.vue";
+import PasswordField from "../PasswordField.vue";
+import PhoneField from "../PhoneField.vue";
+import RatingField from "../RatingField.vue";
+import SelectField from "../SelectField.vue";
+import TableMultiSelectField from "../TableMultiSelectField.vue";
+import TextField from "../TextField.vue";
+import TextareaField from "../TextareaField.vue";
+import TimeField from "../TimeField.vue";
 
 describe("fieldTypes registry", () => {
   it("resolves a registered fieldtype", () => {
